% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_helpers.R
\name{ga_points_shapes_lookup}
\alias{ga_points_shapes_lookup}
\title{Title}
\usage{
ga_points_shapes_lookup(ga_points, shapes = rUKcensus::shapes_lad)
}
\arguments{
\item{population_col}{}
}
\value{

}
\description{
Title
}
