% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_helpers.R
\name{ga_to_polygons}
\alias{ga_to_polygons}
\title{Title}
\usage{
ga_to_polygons(
  ga_points,
  shapes = shapes_regions,
  join_col = "RGN20CD",
  population_col = "mid_2020"
)
}
\arguments{
\item{population_col}{}
}
\value{

}
\description{
Title
}
